source code
A source code is a set of instructions written in a programming language that a computer can understand and execute. It is human-readable and serves as the foundation for creating software applications, scripts, or programs. Source code typically consists of statements, declarations, comments, and other constructs that define the logic and behavior of a program. Overall, source code is the foundation of software development, allowing programmers to create a wide range of applications and systems by expressing their ideas and algorithms in a programming language that computers can understand and execute.